+++ src/domain/accounts/services/tests/tests/use_cases/test_create_account_use_case_impl.rs
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::assert_matches::assert_matches;
+use std::collections::HashSet;
 use std::sync::Arc;
 
 use database_common::NoOpDatabasePlugin;
@@ -38,6 +38,8 @@ use kamu_auth_rebac_services::{
 };
 use messaging_outbox::{MockOutbox, Outbox};
 use odf::AccountName;
+use odf::metadata::DidPkh;
+use pretty_assertions::assert_matches;
 use time_source::SystemTimeSourceDefault;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -47,7 +49,7 @@ const WASYA: &str = "wasya";
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[test_log::test(tokio::test)]
-async fn test_create_account_use_case() {
+async fn test_create_account() {
     let new_account_name_with_email = AccountName::new_unchecked("foo");
     let new_account_email = Email::parse("foo@defined.com").unwrap();
 
@@ -55,17 +57,17 @@ async fn test_create_account_use_case() {
     let new_account_name_without_generated_email = Email::parse("wasya+bar@example.com").unwrap();
 
     let mut mock_outbox = MockOutbox::new();
-    expect_outbox_account_created(
+    expect_outbox_account_created_once(
         &mut mock_outbox,
         AccountDisplayName::from(WASYA),
         format!("{WASYA}@example.com").parse().unwrap(),
     );
-    expect_outbox_account_created(
+    expect_outbox_account_created_once(
         &mut mock_outbox,
         AccountDisplayName::from(new_account_name_with_email.as_str()),
         new_account_email.clone(),
     );
-    expect_outbox_account_created(
+    expect_outbox_account_created_once(
         &mut mock_outbox,
         AccountDisplayName::from(new_account_name_without_email.as_str()),
         new_account_name_without_generated_email.clone(),
@@ -111,6 +113,56 @@ async fn test_create_account_use_case() {
         Ok(account)
             if account.email == new_account_name_without_generated_email
                 && account.account_name == new_account_name_without_email
+    );
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_log::test(tokio::test)]
+async fn test_create_wallet_accounts() {
+    // Addresses got from: https://www.ethereumaddressgenerator.com/
+    let did_pkhs = [
+        "did:pkh:eip155:1:0xbf9a00755BB7d2E904b5F569095220c54E742E07",
+        "did:pkh:eip155:1:0x3F41642f9813eb3be2DBc098dad815c25b9156E8",
+        "did:pkh:eip155:1:0xeCd666A695086c10D8d4AB146D2827842bd15Ef9",
+    ]
+    .into_iter()
+    .map(DidPkh::from_did_str)
+    .map(Result::unwrap)
+    .collect::<HashSet<_>>();
+
+    let mut mock_outbox = MockOutbox::new();
+
+    for did_pkh in &did_pkhs {
+        // E.g. 0xbf9a00755BB7d2E904b5F569095220c54E742E07
+        let wallet_address = did_pkh.wallet_address();
+        expect_outbox_account_created_once(
+            &mut mock_outbox,
+            AccountDisplayName::from(wallet_address),
+            format!("{wallet_address}@example.com").parse().unwrap(),
+        );
+    }
+
+    let harness = CreateAccountUseCaseImplHarness::new(mock_outbox).await;
+    let wallet_addresses_count = did_pkhs.len();
+
+    assert_matches!(
+        harness
+            .use_case
+            .execute_multi_wallet_accounts(did_pkhs.clone())
+            .await,
+        Ok(accounts)
+            if accounts.len() == wallet_addresses_count
+    );
+
+    // Idempotence (but w/o new messages -- controlled by mock_outbox expectations).
+    assert_matches!(
+        harness
+            .use_case
+            .execute_multi_wallet_accounts(did_pkhs)
+            .await,
+        Ok(accounts)
+            if accounts.len() == wallet_addresses_count
     );
 }
 
@@ -166,7 +218,7 @@ impl CreateAccountUseCaseImplHarness {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-fn expect_outbox_account_created(
+fn expect_outbox_account_created_once(
     mock_outbox: &mut MockOutbox,
     expected_display_name: AccountDisplayName,
     expected_email: Email,
@@ -187,6 +239,7 @@ fn expect_outbox_account_created(
             }),
             always(),
         )
+        .times(1)
         .returning(|_, _, _| Ok(()));
 }